requirements.txt
flask==2.3.2
flask-cors==3.0.10
flask_jwt_extended==4.5.2
bcrypt==4.0.1
pyyaml==5.1